Update PlayFabAuthenticationContext.h.ejs
double checking some logic. This should read as:

If d is something AND c is something

But we have to negate it, because passing false is what will stop the code running in the fcheck. We just made it true.

 SO The distirbution property should be applicable so seeing

d == 0 || c == 0

Should still logically also make sense (if both are 0...? then we have what we wanted to avoid)

idk... reading the last one sounds like "if d.length is 0 ||..." that first part is true, so it sounds like it would just skip the second part of the if check if the compiler was optimizing appropriately (it shouldn't need both sides of OR to be evaluated to true if the first argument evaluates to true).

THIS IS WHERE XOR WOULD BE HELPFUL (but that would be rude to have to force developers to read, this logically makes sense and forces the compiler to read how the engineers read. This is more maintainable then the xor).