Update PlayFabAuthenticationContext.h.ejs (#57)
removing ClientAdminSecurityCheck. Devs have been complaining they need this for local test builds. This was intended to prevent shipping a sever security issue, but it has been determined this is more of a hinderance to the development cycle, and devs should already understand admin function calls shouldn't be built into release.