Attempt default enum initialization
in some cases (usually on platforms beyond PC), Unreal needs to "Generate Visual Studio Files" This will trigger a process to go through the code and find the correct starting param values for .h files.

Several devs have ran into this issue, so we will attempt to blindly fix the default variables involved in the error.